% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyenv.r
\name{py_exec}
\alias{py_exec}
\title{Execute Python Code}
\usage{
py_exec(...)
}
\arguments{
\item{...}{Python code to execute.}
}
\value{
A string containing the result, if any.
}
\description{
Execute Python code and return the result.
}
