% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyenv.r
\name{py_get}
\alias{py_get}
\title{Get Python Variable}
\usage{
py_get(x)
}
\arguments{
\item{x}{The Python variable name.}
}
\value{
The value of the Python variable.
}
\description{
Retrieve the value of a Python variable.
}
\examples{
\dontrun{ 
py_exec('a = 5')
py_get('a')
}

}
