% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyenv.r
\name{py_set}
\alias{py_set}
\title{Set Python Variable}
\usage{
py_set(...)
}
\arguments{
\item{...}{A sequence of variables to set. Can be either
existing variables in the R environment or arbitrary 
named variables to create.}
}
\description{
Set a Python variable.
}
\examples{
\dontrun{ 
py_set(a = 5)
py_set(b = 7, c = "foo")
py_set(d = 1:5)
py_set(d = list(1, 2, 3, 4, 5))
}
 
}
