% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyenv.r
\name{py_start}
\alias{py_start}
\title{Start Python}
\usage{
py_start(python.port, python.path)
}
\arguments{
\item{python.port}{The port to use for communication.}

\item{python.path}{The path to the Python executable.}
}
\description{
Start the Python Server.
}
