% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyr6.r
\docType{package}
\name{pysockr-package}
\alias{pysockr-package}
\alias{pysockr}
\title{Socket-based Interface for Python Processes}
\description{
A simple socket-based interface to Python. Provides a basic 
Python server script and R6 class for interacting with a
Python process.
}
